Add files via upload
This version measures/compares the time taken by sterlings approximation and a basic iterative approach to implement the factorial operator. I have tested values up to 1000, and the iterative approach is still faster. More work and research needs to be done to find how or when Sterlings approximation is less computationally intensive than a brute force method. My next addition will be to include the factorial function from the c++ boost library and compare that as well.