Add files via upload
The Colatz conjecture states that any number will reduce to 1 via the following simple formula. If the given number is even, divide it by 2. Otherwise, multiply the number by 3 and add 1. This program displays each step in the process after the formula is applied. My next version will find the amount of steps required for all numbers less than 100 so I can look for patterns. The graphical representation is easily rendered via JavaScript p5.js and is featured in my Beautiful Math repository.